[kernel] bpf: Add bpf_probe_write_user BPF helper to be called in tracers
Message-id: <20180613214317.28803-71-jolsa@redhat.com>
Patchwork-id: 221510
O-Subject: [PATCH RHEL7.6 BZ1311586 070/386] bpf: Add bpf_probe_write_user BPF helper to be called in tracers
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit 96ae52279594470622ff0585621a13e96b700600
Author: Sargun Dhillon <sargun@sargun.me>
Date: Mon Jul 25 05:54:46 2016 -0700

description
===========
This allows user memory to be written to during the course of a kprobe.
It shouldn't be used to implement any kind of security mechanism
because of TOC-TOU attacks, but rather to debug, divert, and
manipulate execution of semi-cooperative processes.

Although it uses probe_kernel_write, we limit the address space
the probe can write into by checking the space with access_ok.
We do this as opposed to calling copy_to_user directly, in order
to avoid sleeping. In addition we ensure the threads's current fs
/ segment is USER_DS and the thread isn't exiting nor a kernel thread.

Given this feature is meant for experiments, and it has a risk of
crashing the system, and running programs, we print a warning on
when a proglet that attempts to use this helper is installed,
along with the pid and process name.

RHEL7
=====
Omitting samples/ hunks.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>