[kernel] bpf: allow access into map value arrays
Message-id: <20180613214317.28803-92-jolsa@redhat.com>
Patchwork-id: 221553
O-Subject: [PATCH RHEL7.6 BZ1311586 091/386] bpf: allow access into map value arrays
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit 484611357c19f9e19ef742ebef4505a07d243cc9
Author: Josef Bacik <jbacik@fb.com>
Date: Wed Sep 28 10:54:32 2016 -0400

description
===========
Suppose you have a map array value that is something like this

struct foo {
	unsigned iter;
	int array[SOME_CONSTANT];
};

You can easily insert this into an array, but you cannot modify the contents of
foo->array[] after the fact.  This is because we have no way to verify we won't
go off the end of the array at verification time.  This patch provides a start
for this work.  We accomplish this by keeping track of a minimum and maximum
value a register could be while we're checking the code.  Then at the time we
try to do an access into a MAP_VALUE we verify that the maximum offset into that
region is a valid access into that memory region.  So in practice, code such as
this

unsigned index = 0;

if (foo->iter >= SOME_CONSTANT)
	foo->iter = index;
else
	index = foo->iter++;
foo->array[index] = bar;

would be allowed, as we can verify that index will always be between 0 and
SOME_CONSTANT-1.  If you wish to use signed values you'll have to have an extra
check to make sure the index isn't less than 0, or do something like index %=
SOME_CONSTANT.

RHEL7
=====
Skip unbackported functions hunks and samples changes.
Small conflict due to missing XDP changes.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>