[kernel] bpf: fix states equal logic for varlen access
Message-id: <20180613214317.28803-110-jolsa@redhat.com>
Patchwork-id: 221592
O-Subject: [PATCH RHEL7.6 BZ1311586 109/386] bpf: fix states equal logic for varlen access
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit e2d2afe15ed452f91797a80dbc0a17838ba03ed4
Author: Josef Bacik <jbacik@fb.com>
Date: Tue Nov 29 12:27:09 2016 -0500

description
===========
If we have a branch that looks something like this

int foo = map->value;
if (condition) {
  foo += blah;
} else {
  foo = bar;
}
map->array[foo] = baz;

We will incorrectly assume that the !condition branch is equal to the condition
branch as the register for foo will be UNKNOWN_VALUE in both cases.  We need to
adjust this logic to only do this if we didn't do a varlen access after we
processed the !condition branch, otherwise we have different ranges and need to
check the other branch as well.

Fixes: 484611357c19 ("bpf: allow access into map value arrays")
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>