[kernel] bpf: fix loading of BPF_MAXINSNS sized programs
Message-id: <20180613214317.28803-114-jolsa@redhat.com>
Patchwork-id: 221552
O-Subject: [PATCH RHEL7.6 BZ1311586 113/386] bpf: fix loading of BPF_MAXINSNS sized programs
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit ef0915cacd04c9e35be5f9d62a4e4b5b4b9bcfd1
Author: Daniel Borkmann <daniel@iogearbox.net>
Date: Wed Dec 7 01:15:44 2016 +0100

description
===========
General assumption is that single program can hold up to BPF_MAXINSNS,
that is, 4096 number of instructions. It is the case with cBPF and
that limit was carried over to eBPF. When recently testing digest, I
noticed that it's actually not possible to feed 4096 instructions
via bpf(2).

The check for > BPF_MAXINSNS was added back then to bpf_check() in
cbd357008604 ("bpf: verifier (add ability to receive verification log)").
However, 09756af46893 ("bpf: expand BPF syscall with program load/unload")
added yet another check that comes before that into bpf_prog_load(),
but this time bails out already in case of >= BPF_MAXINSNS.

Fix it up and perform the check early in bpf_prog_load(), so we can drop
the second one in bpf_check(). It makes sense, because also a 0 insn
program is useless and we don't want to waste any resources doing work
up to bpf_check() point. The existing bpf(2) man page documents E2BIG
as the official error for such cases, so just stick with it as well.

Fixes: 09756af46893 ("bpf: expand BPF syscall with program load/unload")
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>