[kernel] bpf: add prog_digest and expose it via fdinfo/netlink
Message-id: <20180613214317.28803-117-jolsa@redhat.com>
Patchwork-id: 221617
O-Subject: [PATCH RHEL7.6 BZ1311586 116/386] bpf: add prog_digest and expose it via fdinfo/netlink
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit 7bd509e311f408f7a5132fcdde2069af65fa05ae
Author: Daniel Borkmann <daniel@iogearbox.net>
Date: Sun Dec 4 23:19:41 2016 +0100

description
===========
When loading a BPF program via bpf(2), calculate the digest over
the program's instruction stream and store it in struct bpf_prog's
digest member. This is done at a point in time before any instructions
are rewritten by the verifier. Any unstable map file descriptor
number part of the imm field will be zeroed for the hash.

fdinfo example output for progs:

  # cat /proc/1590/fdinfo/5
  pos:          0
  flags:        02000002
  mnt_id:       11
  prog_type:    1
  prog_jited:   1
  prog_digest:  b27e8b06da22707513aa97363dfb11c7c3675d28
  memlock:      4096

When programs are pinned and retrieved by an ELF loader, the loader
can check the program's digest through fdinfo and compare it against
one that was generated over the ELF file's program section to see
if the program needs to be reloaded. Furthermore, this can also be
exposed through other means such as netlink in case of a tc cls/act
dump (or xdp in future), but also through tracepoints or other
facilities to identify the program. Other than that, the digest can
also serve as a base name for the work in progress kallsyms support
of programs. The digest doesn't depend/select the crypto layer, since
we need to keep dependencies to a minimum. iproute2 will get support
for this facility.

RHEL7
=====
Skip networking stuff and changed show_fdinfo prototype
to match RHEL7.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>