[kernel] bpf: pass original insn directly to convert_ctx_access
Message-id: <20180613214317.28803-130-jolsa@redhat.com>
Patchwork-id: 221583
O-Subject: [PATCH RHEL7.6 BZ1311586 129/386] bpf: pass original insn directly to convert_ctx_access
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit 6b8cc1d11ef75c5b9c530b3d0d148f3c2dd25f93
Author: Daniel Borkmann <daniel@iogearbox.net>
Date: Thu Jan 12 11:51:32 2017 +0100

description
===========
Currently, when calling convert_ctx_access() callback for the various
program types, we pass in insn->dst_reg, insn->src_reg, insn->off from
the original instruction. This information is needed to rewrite the
instruction that is based on the user ctx structure into a kernel
representation for the ctx. As we'd like to allow access size beyond
just BPF_W, we'd need also insn->code for that in order to decode the
original access size. Given that, lets just pass insn directly to the
convert_ctx_access() callback and work on that to not clutter the
callback with even more arguments we need to pass when everything is
already contained in insn. So lets go through that once, no functional
change.

RHEL7
=====
Skip the networking code.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>