[kernel] bpf: add a longest prefix match trie map implementation
Message-id: <20180613214317.28803-136-jolsa@redhat.com>
Patchwork-id: 221594
O-Subject: [PATCH RHEL7.6 BZ1311586 135/386] bpf: add a longest prefix match trie map implementation
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit b95a5c4db09bc7c253636cb84dc9b12c577fd5a0
Author: Daniel Mack <daniel@zonque.org>
Date: Sat Jan 21 17:26:11 2017 +0100

description
===========
This trie implements a longest prefix match algorithm that can be used
to match IP addresses to a stored set of ranges.

Internally, data is stored in an unbalanced trie of nodes that has a
maximum height of n, where n is the prefixlen the trie was created
with.

Tries may be created with prefix lengths that are multiples of 8, in
the range from 8 to 2048. The key used for lookup and update operations
is a struct bpf_lpm_trie_key, and the value is a uint64_t.

The code carries more information about the internal implementation.

RHEL7
=====
Skipping testing changes.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>