[kernel] bpf: Fix and simplifications on inline map lookup
Message-id: <20180613214317.28803-156-jolsa@redhat.com>
Patchwork-id: 221614
O-Subject: [PATCH RHEL7.6 BZ1311586 155/386] bpf: Fix and simplifications on inline map lookup
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit fad73a1a35ea61f13607a391aca669caad8c04ca
Author: Martin KaFai Lau <kafai@fb.com>
Date:   Wed Mar 22 10:00:32 2017 -0700

description
===========
Fix in verifier:
For the same bpf_map_lookup_elem() instruction (i.e. "call 1"),
a broken case is "a different type of map could be used for the
same lookup instruction". For example, an array in one case and a
hashmap in another.  We have to resort to the old dynamic call behavior
in this case.  The fix is to check for collision on insn_aux->map_ptr.
If there is collision, don't inline the map lookup.

Please see the "do_reg_lookup()" in test_map_in_map_kern.c in the later
patch for how-to trigger the above case.

Simplifications on array_map_gen_lookup():
1. Calculate elem_size from map->value_size.  It removes the
   need for 'struct bpf_array' which makes the later map-in-map
   implementation easier.
2. Remove the 'elem_size == 1' test

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>