[kernel] bpf: map_get_next_key to return first key on NULL
Message-id: <20180613214317.28803-167-jolsa@redhat.com>
Patchwork-id: 221679
O-Subject: [PATCH RHEL7.6 BZ1311586 166/386] bpf: map_get_next_key to return first key on NULL
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit 8fe45924387be6b5c1be59a7eb330790c61d5d10
Author: Teng Qin <qinteng@fb.com>
Date:   Mon Apr 24 19:00:37 2017 -0700

description
===========
When iterating through a map, we need to find a key that does not exist
in the map so map_get_next_key will give us the first key of the map.
This often requires a lot of guessing in production systems.

This patch makes map_get_next_key return the first key when the key
pointer in the parameter is NULL.

RHEL7
=====
Skip tracepoint and net changes.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>