[kernel] bpf: Add BPF_OBJ_GET_INFO_BY_FD
Message-id: <20180613214317.28803-177-jolsa@redhat.com>
Patchwork-id: 221680
O-Subject: [PATCH RHEL7.6 BZ1311586 176/386] bpf: Add BPF_OBJ_GET_INFO_BY_FD
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit 1e270976908686ec25fb91b8a34145be54137976
Author: Martin KaFai Lau <kafai@fb.com>
Date: Mon Jun 5 12:15:52 2017 -0700

description
===========
A single BPF_OBJ_GET_INFO_BY_FD cmd is used to obtain the info
for both bpf_prog and bpf_map.  The kernel can figure out the
fd is associated with a bpf_prog or bpf_map.

The suggested struct bpf_prog_info and struct bpf_map_info are
not meant to be a complete list and it is not the goal of this patch.
New fields can be added in the future patch.

The focus of this patch is to create the interface,
BPF_OBJ_GET_INFO_BY_FD cmd for exposing the bpf_prog's and
bpf_map's info.

The obj's info, which will be extended (and get bigger) over time, is
separated from the bpf_attr to avoid bloating the bpf_attr.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>