[kernel] bpf: devmap: remove unnecessary value size check
Message-id: <20180613214317.28803-229-jolsa@redhat.com>
Patchwork-id: 221717
O-Subject: [PATCH RHEL7.6 BZ1311586 228/386] bpf: devmap: remove unnecessary value size check
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit cf9d01405925e3f8144c99d7bf7b184449794066
Author: John Fastabend <john.fastabend@gmail.com>
Date: Tue Aug 15 23:35:12 2017 -0700

description
===========
In the devmap alloc map logic we check to ensure that the sizeof the
values are not greater than KMALLOC_MAX_SIZE. But, in the dev map case
we ensure the value size is 4bytes earlier in the function because all
values should be netdev ifindex values.

The second check is harmless but is not needed so remove it.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>