[kernel] bpf: fix out-of-bounds access warning in bpf_check
Message-id: <20180613214317.28803-276-jolsa@redhat.com>
Patchwork-id: 221697
O-Subject: [PATCH RHEL7.6 BZ1311586 275/386] bpf: fix out-of-bounds access warning in bpf_check
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit eba0c929d1d0f16c4b03628b7bf8ce363b9e5c9a
Author: Arnd Bergmann <arnd@arndb.de>
Date: Thu Nov 2 12:05:52 2017 +0100

description
===========
The bpf_verifer_ops array is generated dynamically and may be
empty depending on configuration, which then causes an out
of bounds access:

kernel/bpf/verifier.c: In function 'bpf_check':
kernel/bpf/verifier.c:4320:29: error: array subscript is above array bounds [-Werror=array-bounds]

This adds a check to the start of the function as a workaround.
I would assume that the function is never called in that configuration,
so the warning is probably harmless.

Fixes: 00176a34d9e2 ("bpf: remove the verifier ops from program structure")
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>