[kernel] bpf: offload: add infrastructure for loading programs for a specific netdev
Message-id: <20180613214317.28803-278-jolsa@redhat.com>
Patchwork-id: 221700
O-Subject: [PATCH RHEL7.6 BZ1311586 277/386] bpf: offload: add infrastructure for loading programs for a specific netdev
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit ab3f0063c48c26c927851b6767824e35a716d878
Author: Jakub Kicinski <jakub.kicinski@netronome.com>
Date: Fri Nov 3 13:56:17 2017 -0700

description
===========
The fact that we don't know which device the program is going
to be used on is quite limiting in current eBPF infrastructure.
We have to reverse or limit the changes which kernel makes to
the loaded bytecode if we want it to be offloaded to a networking
device.  We also have to invent new APIs for debugging and
troubleshooting support.

Make it possible to load programs for a specific netdev.  This
helps us to bring the debug information closer to the core
eBPF infrastructure (e.g. we will be able to reuse the verifer
log in device JIT).  It allows device JITs to perform translation
on the original bytecode.

__bpf_prog_get() when called to get a reference for an attachment
point will now refuse to give it if program has a device assigned.
Following patches will add a version of that function which passes
the expected netdev in. @type argument in __bpf_prog_get() is
renamed to attach_type to make it clearer that it's only set on
attachment.

All calls to ndo_bpf are protected by rtnl, only verifier callbacks
are not.  We need a wait queue to make sure netdev doesn't get
destroyed while verifier is still running and calling its driver.

RHEL7
=====
Adding only bpf_prog_is_dev_bound stub.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>