[x86] bpf, x86: small optimization in alu ops with imm
Message-id: <20180613214317.28803-335-jolsa@redhat.com>
Patchwork-id: 221772
O-Subject: [PATCH RHEL7.6 BZ1311586 334/386] bpf, x86: small optimization in alu ops with imm
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit de0a444ddae61b77683c27cd390e518e33858d20
Author: Daniel Borkmann <daniel@iogearbox.net>
Date: Sat Jan 20 01:24:35 2018 +0100

description
===========
For the BPF_REG_0 (BPF_REG_A in cBPF, respectively), we can use
the short form of the opcode as dst mapping is on eax/rax and
thus save a byte per such operation. Added to add/sub/and/or/xor
for 32/64 bit when K immediate is used. There may be more such
low-hanging fruit to add in future as well.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>