[x86] bpf, x64: fix memleak when not converging on calls
Message-id: <20180613214317.28803-359-jolsa@redhat.com>
Patchwork-id: 221825
O-Subject: [PATCH RHEL7.6 BZ1311586 358/386] bpf, x64: fix memleak when not converging on calls
Bugzilla: 1311586
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1311586
https://bugzilla.redhat.com/show_bug.cgi?id=1311586

upstream
========
commit 39f56ca945af86112753646316c4c92dcd4acd82
Author: Daniel Borkmann <daniel@iogearbox.net>
Date: Wed May 2 20:12:23 2018 +0200

description
===========
The JIT logic in jit_subprogs() is as follows: for all subprogs we
allocate a bpf_prog_alloc(), populate it (prog->is_func = 1 here),
and pass it to bpf_int_jit_compile(). If a failure occurred during
JIT and prog->jited is not set, then we bail out from attempting to
JIT the whole program, and punt to the interpreter instead. In case
JITing went successful, we fixup BPF call offsets and do another
pass to bpf_int_jit_compile() (extra_pass is true at that point) to
complete JITing calls. Given that requires to pass JIT context around
addrs and jit_data from x86 JIT are freed in the extra_pass in
bpf_int_jit_compile() when calls are involved (if not, they can
be freed immediately). However, if in the original pass, the JIT
image didn't converge then we leak addrs and jit_data since image
itself is NULL, the prog->is_func is set and extra_pass is false
in that case, meaning both will become unreachable and are never
cleaned up, therefore we need to free as well on !image. Only x64
JIT is affected.

Fixes: 1c2a088a6626 ("bpf: x64: add JIT support for multi-function programs")

RHEL7
=====
Making changes in trace_bpf_jit_comp.c.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>