[kernel] redhat: fix kABI for dynamic ftrace on powerpc
Message-id: <e8ee47f832ce0130979ced3724eaa088a463cdd5.1529509954.git.jpoimboe@redhat.com>
Patchwork-id: 222640
O-Subject: [RHEL7.6 PATCH v2 26/29] redhat: fix kABI for dynamic ftrace on powerpc
Bugzilla: 1578960
RH-Acked-by: Artem Savkov <asavkov@redhat.com>
RH-Acked-by: Joe Lawrence <joe.lawrence@redhat.com>

This patch is RHEL-only.

In order to enable dynamic ftrace, some additional fields need to be
added to struct module and task_struct.  This needs to be done carefully
so as not to break kABI.

Most of this path has already been trailblazed by enabling ftrace on
s390, so adjust those ifdefs to support ppc64le as well.  Also reverse
the ifdef conditions such that arm64 will also get the same benefit if
we ever enable dynamic ftrace there.

Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>