[scripts] redhat: fix kABI for -mprofile-kernel
Message-id: <9e962213100d66bbd1ee4d14c608a395304bf2d2.1529509954.git.jpoimboe@redhat.com>
Patchwork-id: 222642
O-Subject: [RHEL7.6 PATCH v2 27/29] redhat: fix kABI for -mprofile-kernel
Bugzilla: 1578960
RH-Acked-by: Artem Savkov <asavkov@redhat.com>
RH-Acked-by: Joe Lawrence <joe.lawrence@redhat.com>

This patch is RHEL-only.

The ppc64le kernel is soon going to be built with -mprofile-kernel.  If
an old module is loaded which wasn't compiled with this flag, it will
produce a warning, as described in the following upstream commit:

  43e24e82f352 ("powerpc/modules: If mprofile-kernel is enabled add it to vermagic")

The upstream solution to this problem (as implemented in the above
commit) is to prevent such an old module from loading.  That doesn't
work for RHEL because it would break kABI.

So instead, just prevent ftrace from processing the module's mcount
locs.  This will effectively disable ftrace for all of the module's
functions.  This isn't a regression because before RHEL 7.6 we didn't
have *any* ftrace support for ppc64le.

On the other hand, modules built against RHEL 7.6+ will be built with
-mprofile-kernel and will work fine.

Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>