[md] ensure md devices are freed before module is unloaded
Message-id: <1530561984-30263-19-git-send-email-ncroxon@redhat.com>
Patchwork-id: 223357
O-Subject: [RHEL7.6 V3 PATCH 018/136] md: ensure md devices are freed before module is unloaded.
Bugzilla: 1494474
RH-Acked-by: Heinz Mauelshagen <heinzm@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>

From: NeilBrown <neilb@suse.com>

Commit: cbd199837750 ("md: Fix unfortunate interaction with evms")
change mddev_put() so that it would not destroy an md device while
->ctime was non-zero.

Unfortunately, we didn't make sure to clear ->ctime when unloading
the module, so it is possible for an md device to remain after
module unload.  An attempt to open such a device will trigger
an invalid memory reference in:
  get_gendisk -> kobj_lookup -> exact_lock -> get_disk

when tring to access disk->fops, which was in the module that has
been removed.

So ensure we clear ->ctime in md_exit(), and explain how that is useful,
as it isn't immediately obvious when looking at the code.

Fixes: cbd199837750 ("md: Fix unfortunate interaction with evms")
Tested-by: Guoqing Jiang <gqjiang@suse.com>
Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Shaohua Li <shli@fb.com>
(cherry picked from commit 9356863c9409efc79029c01a85d015efae977e69)
Signed-off-by: Nigel Croxon <ncroxon@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>