[md] raid5: remove over-loading of ->bi_phys_segments
Message-id: <1530561984-30263-43-git-send-email-ncroxon@redhat.com>
Patchwork-id: 223331
O-Subject: [RHEL7.6 V3 PATCH 042/136] md/raid5: remove over-loading of ->bi_phys_segments.
Bugzilla: 1494474
RH-Acked-by: Heinz Mauelshagen <heinzm@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>

From: NeilBrown <neilb@suse.com>

When a read request, which bypassed the cache, fails, we need to retry
it through the cache.
This involves attaching it to a sequence of stripe_heads, and it may not
be possible to get all the stripe_heads we need at once.
We do what we can, and record how far we got in ->bi_phys_segments so
we can pick up again later.

There is only ever one bio which may have a non-zero offset stored in
->bi_phys_segments, the one that is either active in the single thread
which calls retry_aligned_read(), or is in conf->retry_read_aligned
waiting for retry_aligned_read() to be called again.

So we only need to store one offset value.  This can be in a local
variable passed between remove_bio_from_retry() and
retry_aligned_read(), or in the r5conf structure next to the
->retry_read_aligned pointer.

Storing it there allows the last usage of ->bi_phys_segments to be
removed from md/raid5.c.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Shaohua Li <shli@fb.com>
(cherry picked from commit 0472a42ba1f89ec85f070c731f4440d7cc38c44c)
Signed-off-by: Xiao Ni <xni@redhat.com>
Signed-off-by: Nigel Croxon <ncroxon@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>