[md] raid6: Fix anomily when recovering a single device in RAID6
Message-id: <1530561984-30263-61-git-send-email-ncroxon@redhat.com>
Patchwork-id: 223385
O-Subject: [RHEL7.6 V3 PATCH 060/136] md/raid6: Fix anomily when recovering a single device in RAID6.
Bugzilla: 1494474
RH-Acked-by: Heinz Mauelshagen <heinzm@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>

From: NeilBrown <neilb@suse.com>

When recoverying a single missing/failed device in a RAID6,
those stripes where the Q block is on the missing device are
handled a bit differently.  In these cases it is easy to
check that the P block is correct, so we do.  This results
in the P block be destroy.  Consequently the P block needs
to be read a second time in order to compute Q.  This causes
lots of seeks and hurts performance.

It shouldn't be necessary to re-read P as it can be computed
from the DATA.  But we only compute blocks on missing
devices, since c337869d9501 ("md: do not compute parity
unless it is on a failed drive").

So relax the change made in that commit to allow computing
of the P block in a RAID6 which it is the only missing that
block.

This makes RAID6 recovery run much faster as the disk just
"before" the recovering device is no longer seeking
back-and-forth.

Reported-by-tested-by: Brad Campbell <lists2009@fnarfbargle.com>
Reviewed-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Shaohua Li <shli@fb.com>
(cherry picked from commit 7471fb77ce4dc4cb81291189947fcdf621a97987)
Signed-off-by: Nigel Croxon <ncroxon@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>