[md] raid1: abort delayed writes when device fails
Message-id: <1530561984-30263-66-git-send-email-ncroxon@redhat.com>
Patchwork-id: 223332
O-Subject: [RHEL7.6 V3 PATCH 065/136] md/raid1: abort delayed writes when device fails.
Bugzilla: 1494474
RH-Acked-by: Heinz Mauelshagen <heinzm@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>

From: NeilBrown <neilb@suse.com>

When writing to an array with a bitmap enabled, the writes are grouped
in batches which are preceded by an update to the bitmap.

It is quite likely if that a drive develops a problem which is not
media related, that the bitmap write will be the first to report an
error and cause the device to be marked faulty (as the bitmap write is
at the start of a batch).

In this case, there is point submiting the subsequent writes to the
failed device - that just wastes times.

So re-check the Faulty state of a device before submitting a
delayed write.

This requires that we keep the 'rdev', rather than the 'bdev' in the
bio, then swap in the bdev just before final submission.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Shaohua Li <shli@fb.com>
(cherry picked from commit 5e2c7a3611977b69ae0531e8fbdeab5dad17925a)
Signed-off-by: Nigel Croxon <ncroxon@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>

Conflicts:
	drivers/md/raid1.c
RHEL7 does not support bio member bi_error.