[md] report sector of stripes with check mismatches
Message-id: <1530561984-30263-79-git-send-email-ncroxon@redhat.com>
Patchwork-id: 223392
O-Subject: [RHEL7.6 V3 PATCH 078/136] md: report sector of stripes with check mismatches
Bugzilla: 1494474
RH-Acked-by: Heinz Mauelshagen <heinzm@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>

From: Nix <nix@esperi.org.uk>

This makes it possible, with appropriate filesystem support, for a
sysadmin to tell what is affected by the mismatch, and whether
it should be ignored (if it's inside a swap partition, for
instance).

We ratelimit to prevent log flooding: if there are so many
mismatches that ratelimiting is necessary, the individual messages
are relatively unlikely to be important (either the machine is
swapping like crazy or something is very wrong with the disk).

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Signed-off-by: Shaohua Li <shli@fb.com>
(cherry picked from commit e153903686deac9ea8c499b49516126f1a844fdb)
Signed-off-by: Nigel Croxon <ncroxon@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>