[security] bpf: Add LSM hooks for bpf object related syscall
Message-id: <20180705072910.5792-3-jolsa@redhat.com>
Patchwork-id: 223498
O-Subject: [PATCH RHEL7.6 BZ1593027 2/3] security: bpf: Add LSM hooks for bpf object related syscall
Bugzilla: 1593027
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>

Bugzilla: 1593027
https://bugzilla.redhat.com/show_bug.cgi?id=1593027

upstream
========
commit afdb09c720b62b8090584c11151d856df330e57d
Author: Chenbo Feng <fengc@google.com>
Date: Wed Oct 18 13:00:24 2017 -0700

description
===========
Introduce several LSM hooks for the syscalls that will allow the
userspace to access to eBPF object such as eBPF programs and eBPF maps.
The security check is aimed to enforce a per object security protection
for eBPF object so only processes with the right priviliges can
read/write to a specific map or use a specific eBPF program. Besides
that, a general security hook is added before the multiplexer of bpf
syscall to check the cmd and the attribute used for the command. The
actual security module can decide which command need to be checked and
how the cmd should be checked.

RHEL7
=====
This is the leftover of the previous backport of this patch:
  b7d9a2dfb8ba [security] bpf: Add LSM hooks for bpf object related syscall
which kept the hooks disabled.

Enabling bpf hooks now by calling security_ops bpf callbacks, which
is different from upstream way of calling call_int_hook helper.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>