[security] selinux: bpf: Add addtional check for bpf object file receive
Message-id: <20180705072910.5792-4-jolsa@redhat.com>
Patchwork-id: 223495
O-Subject: [PATCH RHEL7.6 BZ1593027 3/3] selinux: bpf: Add addtional check for bpf object file receive
Bugzilla: 1593027
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>

Bugzilla: 1593027
https://bugzilla.redhat.com/show_bug.cgi?id=1593027

upstream
========
commit f66e448cfda021b0bcd884f26709796fe19c7cc1
Author: Chenbo Feng <fengc@google.com>
Date:   Wed Oct 18 13:00:26 2017 -0700

description
===========
Introduce a bpf object related check when sending and receiving files
through unix domain socket as well as binder. It checks if the receiving
process have privilege to read/write the bpf map or use the bpf program.
This check is necessary because the bpf maps and programs are using a
anonymous inode as their shared inode so the normal way of checking the
files and sockets when passing between processes cannot work properly on
eBPF object. This check only works when the BPF_SYSCALL is configured.

RHEL7
=====
Skipping the change in selinux_binder_transfer_file,
because it's not backported in RHEL7.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>