[kernel] sched/wait: Improve __var_waitqueue() code generation
Message-id: <20180705193835.25057-32-jmoyer@redhat.com>
Patchwork-id: 223549
O-Subject: [RHEL7.6 PATCH 31/31] sched/wait: Improve __var_waitqueue() code generation
Bugzilla: 1505291
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

From: Peter Zijlstra <peterz@infradead.org>

BZ: 1505291
Brew: 17007175

Conflicts: RHEL still has the wait functions in kernel/wait.c.
The comment below about changing the hash function to not suck is
referring to the following commits, which we don't have:

689de1d6ca9: Minimal fix-up of bad hashing behavior of hash_64()
ef703f49a6c: Eliminate bad hash multipliers from hash_32() and  hash_64()

I'm pulling this patch in not as a clean-up, but rather as a correctness
fix.  I'm not sure the upstream version of the code ever actually
worked with the .config that we use for RHEL 7.  Basically, the
bit_waitqueue() call would end up trying to get to the page struct that
describes a page struct.  Larry told me that's one level of indirection
too deep to make sense.  This patch just gets rid of that bad behavior.

Without this patch, we'd get a panic in bit_waitqueue.  With the patch
applied, that's gone.

commit b3fc5c9bb373661224906bf434c09ca0de032e82
Author: Peter Zijlstra <peterz@infradead.org>
Date:   Thu Mar 15 11:49:36 2018 +0100

    sched/wait: Improve __var_waitqueue() code generation

    Since we fixed hash_64() to not suck there is no need to play games to
    attempt to improve the hash value on 64-bit.

    Also, since we don't use the bit value for the variables, use hash_ptr()
    directly.

    No change in functionality.

    Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
    Cc: George Spelvin <linux@sciencehorizons.net>
    Cc: Linus Torvalds <torvalds@linux-foundation.org>
    Cc: Mike Galbraith <efault@gmx.de>
    Cc: Peter Zijlstra <peterz@infradead.org>
    Cc: Thomas Gleixner <tglx@linutronix.de>
    Cc: linux-kernel@vger.kernel.org
    Signed-off-by: Ingo Molnar <mingo@kernel.org>

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>