[tools] bpf: add test for the verifier equal logic bug
Message-id: <20180709105913.17980-3-jolsa@redhat.com>
Patchwork-id: 224115
O-Subject: [PATCH RHEL7.6 BZ1593026 02/50] bpf: add test for the verifier equal logic bug
Bugzilla: 1593026
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Bugzilla: 1593026
https://bugzilla.redhat.com/show_bug.cgi?id=1593026

upstream
========
commit e95489010b7abf6d360dfb7d88a41dfa9e0cc802
Author: Josef Bacik <jbacik@fb.com>
Date: Tue Nov 29 12:35:19 2016 -0500

description
===========
This is a test to verify that

bpf: fix states equal logic for varlen access

actually fixed the problem.  The problem was if the register we added to our map
register was UNKNOWN in both the false and true branches and the only thing that
changed was the range then we'd incorrectly assume that the true branch was
valid, which it really wasnt.  This tests this case and properly fails without
my fix in place and passes with it in place.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>