[kernel] bpf: comment why dots in filenames under BPF virtual FS are not allowed
Message-id: <20180709105913.17980-10-jolsa@redhat.com>
Patchwork-id: 224116
O-Subject: [PATCH RHEL7.6 BZ1593026 09/50] bpf: comment why dots in filenames under BPF virtual FS are not allowed
Bugzilla: 1593026
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Bugzilla: 1593026
https://bugzilla.redhat.com/show_bug.cgi?id=1593026

upstream
========
commit 6d8cb045cde681e64a5ed80a2ab70be831a7f9b0
Author: Quentin Monnet <quentin.monnet@netronome.com>
Date: Thu Mar 8 23:46:33 2018 -0800

description
===========
When pinning a file under the BPF virtual file system (traditionally
/sys/fs/bpf), using a dot in the name of the location to pin at is not
allowed. For example, trying to pin at "/sys/fs/bpf/foo.bar" will be
rejected with -EPERM.

This check was introduced at the same time as the BPF file system
itself, with commit b2197755b263 ("bpf: add support for persistent
maps/progs"). At this time, it was checked in a function called
"bpf_dname_reserved()", which made clear that using a dot was reserved
for future extensions.

This function disappeared and the check was moved elsewhere with commit
0c93b7d85d40 ("bpf: reject invalid names right in ->lookup()"), and the
meaning of the dot ban was lost.

The present commit simply adds a comment in the source to explain to the
reader that the usage of dots is reserved for future usage.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>