[kernel] tracing: Fix possible out of bounds memory access when parsing enums
Message-id: <20180709105913.17980-48-jolsa@redhat.com>
Patchwork-id: 224096
O-Subject: [PATCH RHEL7.6 BZ1593026 47/50] tracing: Fix possible out of bounds memory access when parsing enums
Bugzilla: 1593026
RH-Acked-by: Eugene Syromiatnikov <esyr@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Benc <jbenc@redhat.com>

Bugzilla: 1593026
https://bugzilla.redhat.com/show_bug.cgi?id=1593026

upstream
========
commit 3193899d4dd54056f8c2e0b1e40dd6e2f0009f28
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date:   Fri Apr 17 10:27:57 2015 -0400

description
===========
The code that replaces the enum names with the enum values in the
tracepoints' format files could possible miss the end of string nul
character. This was caused by processing things like backslashes, quotes
and other tokens. After processing the tokens, a check for the nul
character needed to be done before continuing the loop, because the loop
incremented the pointer before doing the check, which could bypass the nul
character.

Link: http://lkml.kernel.org/r/552E661D.5060502@oracle.com
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>