[kernel] rhel: switch get_fo_extend over to using the registered ops
Message-id: <x49va9man7l.fsf_-_@segfault.boston.devel.redhat.com>
Patchwork-id: 224302
O-Subject: [RHEL7.6 PATCH, v3 52/55] RHEL: switch get_fo_extend over to using the registered ops
Bugzilla: 1598279 1520492 1568236
RH-Acked-by: Ming Lei <ming.lei@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

BZ: 1568236
Brew: 16933005

RHEL 7 commit db0c5ce4b6288 ("[fs] red hat kabi: Added flag signifying
the use of file_operations_extend structure") added a flag to the
file_system_type to signal the presence of an extended file_operations
struct.  This mostly works for today's users, but only by accident.
The way we get to a file_system_type from a struct file is through the
super block:

  file_inode(fp)->i_sb->s_type->fs_flags

The superblock for an inode has nothing to do with the file_operations
struct associated with the file.  The latter can be reassigned during
open, and will be different for different types of files (for example:
character devices, block devices, etc).  What we really want is a way
to register a file_operations_extend structure.  The previous patch in
this series provided a way to do that.  This patch now switches the
get_fo_extend helper to use lookup_fo_extend, which searches a hash
table of registered structures.

Note that we use file_operations and file_operations_extend pointers
interchangeably.  The file_operations embedded in the extended
structure was always the first member.  This means that the
to_fo_extend macro was a glorified cast.

Note also that this fixes commit 73b438b3fb85 ("[fs] ext4: Support for
synchronous DAX faults").  That patch added FS_HAS_FO_EXTEND to the
ext2_fs_type instead of ext4_fs_type.

The next patch in the series will get rid of the bogus FS_HAS_FO_EXTEND
flag set in the various file_system_type structures.

In case it's not obvious, this is a RHEL-only patch.

Fixes: 73b438b3fb85 "[fs] ext4: Support for synchronous DAX faults"
Fixes: 383d7cf7efd4 "[fs] btrfs: add .copy_file_range file operation"

---
v3: Fix up context differences due to an updated patch 50/55.
    Add a fixes tag for btrfs.

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>