[nvdimm] fix btt partition scanning on boot
Message-id: <20180727204336.27083-2-jmoyer@redhat.com>
Patchwork-id: 225583
O-Subject: [RHEL7.6 PATCH,v2 1/2] [RHEL] fix btt partition scanning on boot
Bugzilla: 1600496
RH-Acked-by: David Milburn <dmilburn@redhat.com>
RH-Acked-by: Ming Lei <ming.lei@redhat.com>

BZ: 1600496
Brew: 17391035

Partitions on a btt device do not show up on boot.  The reason is that
the driver does a set_capacity(disk, 0) before calling add_disk.  This
prevents the partition code from reading the partition table.  The
reason *that's* done is because our bio integrity setup code needs the
gendisk to be added before it is called.  However, add_disk can't
perform I/O (to do the partition table read) until we know what size
I/O the device expects (which is filled in by the integrity setup
code).

Upstream fixed this by making the protection information a part of the
gendisk (and then again moving it to the request_queue), thus removing
this allocation step that got us into the chicken and egg situation.
I tried pulling that series back, and it was a real chore.[1]
Ultimately, I decided not to pursue that route.

The second fix I tried was to initiate a partition table scan from
userspace, using a udev rule.  That works well, except if you are
booting from a btt device.  For boot, you need that rule added to the
initramfs.  Doing that requires creating a dracut module, which tests
for the presence of the device, and then plunks the udev rule down.
The dracut package owner thought it would be better to include that
module in the ndctl package, rather than dracut.  The problem there is
that the ndctl package isn't installed by default, and I don't know of
a way to have anaconda automatically install a package based on the
hardware detected on a system.  I also don't think we'd want to do
that, anyway.  This is just piling on the band-aids.

So, the last solution I am pursuing is to just remove the integrity
support from the btt driver (and the nd_blk driver, for which there is
currently no hardware).  The integrity support is basically a lie,
anyway.  I talked to Dan Williams about it, and his response was,
"kill it with fire."

'Nuff said.

In case it wasn't obvious, this is a RHEL-only patch.

[1] http://post-office.corp.redhat.com/archives/rhkernel-list/2018-June/msg02344.html

v2: Restrict sector sizes to 512 and 4096.  Mark devices with PI sector
    sizes read-only.
Signed-off-by: Jeff Moyer <jmoyer@redhat.com>

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>