[netdrv] ibmvnic: Revise RX/TX queue error messages
Message-id: <20180731085721.22240.4628.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 225663
O-Subject: [PATCH RHEL7.6 BZ1609814 1/2] ibmvnic: Revise RX/TX queue error messages
Bugzilla: 1609814
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1609814

Description:
------------
During a device failover, there may be latency between the loss
of the current backing device and a notification from firmware that
a failover has occurred. This latency can result in a large amount of
error printouts as firmware returns outgoing traffic with a generic
error code. These are not necessarily errors in this case as the
firmware is busy swapping in a new backing adapter and is not ready
to send packets yet. This patch reclassifies those error codes as
warnings with an explanation that a failover may be pending. All
other return codes will be considered errors.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Upstream Status:
----------------
commit 2d14d3795294d42aacc278948984a480569bcc23

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=2d14d3795294d42aacc278948984a480569bcc23

----------------
Steve Best

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>