[netdrv] ibmvnic: Fix error recovery on login failure
Message-id: <20180731085729.22240.86231.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 225664
O-Subject: [PATCH RHEL7.6 BZ1609814 2/2] ibmvnic: Fix error recovery on login failure
Bugzilla: 1609814
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1609814

Description:
------------
Testing has uncovered a failure case that is not handled properly. In the
event that a login fails and we are not able to recover on the spot, we
return 0 from do_reset, preventing any error recovery code from being
triggered.  Additionally, the state is set to "probed" meaning that when we
are able to trigger the error recovery, the driver always comes up in the
probed state. To handle the case properly, we need to return a failure code
here and set the adapter state to the state that we entered the reset in
indicating the state that we would like to come out of the recovery reset
in.

Signed-off-by: John Allen <jallen@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Upstream Status:
----------------
commit 3578a7ecb69920efc3885dbd610e98c00dbdf5db

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=3578a7ecb69920efc3885dbd610e98c00dbdf5db

----------------
Steve Best

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>