[redhat] spec: compute content hash for kernel-headers
Message-id: <9a24869f34077a859e5ed3b9dd665255334e0985.1533747780.git.aquini@redhat.com>
Patchwork-id: 226351
O-Subject: [RHEL7 v4 PATCH] redhat: spec: compute content hash for kernel-headers
Bugzilla: 1613000
RH-Acked-by: Jan Stancek <jstancek@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Bruno Eduardo de Oliveira Meneguele <bmeneg@redhat.com>

RHEL-only, not upstream

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1613000

In order to avoid potential update storms to consumers of kernel-headers
when the newer kernel RPMs get released for CVEs but exported header
files do not change we're going to start providing a content-hash
for the installed headers within every kernel-headers-V-R RPM so
it can be compared in order to determine if an update is really
required, or not, for this particular kernel RPM subpackage.

This change will specially avoid Container churn as described
at the BZ as well as it will save us from an even more troublesome
decoupling of kernel-headers from kernel RPM, as far as RHEL is
concerned.

Signed-off-by: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>