[x86] kvm: mitigation for L1 cache terminal fault vulnerabilities
Message-id: <1533921961-30116-2-git-send-email-crecklin@redhat.com>
Patchwork-id: 1741
O-Subject: [RH Confidential: Omega] [RHEL 7.6 CVE-2018-3620 BZ1593384 v5 01/68] kvm: mitigation for L1 cache terminal fault vulnerabilities
Bugzilla: 1593384
CVE: CVE-2018-3620
RH-Acked-by: Jan Stancek <jstancek@redhat.com>
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>

Message-id: <150350e598d9d993089d7f4d6c68fa690934b4a4.1530497845.git.jpoimboe@redhat.com>
Patchwork-id: 862
O-Subject: [RH Confidential: Omega] [RHEL7.5.z PATCH 01/38] kvm: x86: mitigation for L1 cache terminal fault vulnerabilities

From: Paolo Bonzini <pbonzini@redhat.com>

This patch adds two mitigation modes for CVE-2018-3620, aka L1 terminal
fault.  The two modes are "vmexit_l1d_flush=1" and "vmexit_l1d_flush=2".

"vmexit_l1d_flush=2" is simply doing an L1 cache flush on every vmexit.
"vmexit_l1d_flush=1" is trying to avoid so on vmexits that are "confined"
in the kind of code that they execute.  The idea is based on Intel's
patches, but the final list of "confined" vmexits isn't quite.

Notably, L1 cache flushes are performed on EPT violations (which are
basically KVM-level page faults), vmexits that involve the emulator,
and on every KVM_RUN invocation (so each userspace exit).  However,
most vmexits are considered safe.  I singled out the emulator because
it may be a good target for other speculative execution-based threats,
and the MMU because it can bring host page tables in the L1 cache.

The mitigation does not in any way try to do anything about hyperthreading;
it is possible for a sibling thread to read data from the cache during a
vmexit, before the host completes the flush, or to read data from the cache
while a sibling runs.  This part of the work is not ready yet.

For now I'm leaving the default to "vmexit_l1d_flush=2", since this is an emergency
kernel, but I'm not sure it's the best setting.  The cost is, depending
on the processor, 30-50% more expensive vmexits---independent of whether
microcode or the generic flush code are used.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>