[x86] l1tf: Protect swap entries against L1TF
Message-id: <1533921961-30116-7-git-send-email-crecklin@redhat.com>
Patchwork-id: 1731
O-Subject: [RH Confidential: Omega] [RHEL 7.6 CVE-2018-3620 BZ1593384 v5 06/68] l1tf: Protect swap entries against L1TF
Bugzilla: 1593384
CVE: CVE-2018-3620
RH-Acked-by: Jan Stancek <jstancek@redhat.com>
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>

Message-id: <5a60405acb0b05280feed4eab53ed68f6c7d5199.1530497845.git.jpoimboe@redhat.com>
Patchwork-id: 883
O-Subject: [RH Confidential: Omega] [RHEL7.5.z PATCH 06/38] x86, l1tf: Protect swap entries against L1TF

RHEL conflicts:
- missing eee4818baac0 ("mm: x86: move _PAGE_SWP_SOFT_DIRTY from bit 7
  to bit 1")

With L1 terminal fault the CPU speculates into unmapped PTEs, and
resulting side effects allow to read the memory the PTE is pointing
too, if its values are still in the L1 cache.

For swapped out pages Linux uses unmapped PTEs and stores a swap entry
into them.

We need to make sure the swap entry is not pointing to valid memory,
which requires setting higher bits (between bit 36 and bit 45) that
are inside the CPUs physical address space, but outside any real
memory.

To do this we invert the offset to make sure the higher bits are always
set, as long as the swap file is not too big.

Here's a patch that switches the order of "type" and
"offset" in the x86-64 encoding, in addition to doing the binary 'not' on
the offset.

That means that now the offset is bits 9-58 in the page table, and that
the type is in the bits that hardware generally doesn't care about.

That, in turn, means that if you have a desktop chip with only 40 bits of
physical addressing, now that the offset starts at bit 9, you still have
to have 30 bits of offset actually *in use* until bit 39 ends up being
clear.

So that's 4 terabyte of swap space (because the offset is counted in
pages, so 30 bits of offset is 42 bits of actual coverage). With bigger
physical addressing, that obviously grows further, until you hit the limit
of the offset (at 50 bits of offset - 62 bits of actual swap file
coverage).

Note there is no workaround for 32bit !PAE, or on systems which
have more than MAX_PA/2 memory. The later case is very unlikely
to happen on real systems.

[updated description and minor tweaks by AK]

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Andi Kleen <ak@linux.intel.com>
Tested-by: Andi Kleen <ak@linux.intel.com>
Acked-by: Michal Hocko <mhocko@suse.com>
Acked-by: Vlastimil Babka <vbabka@suse.cz>
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>