[kernel] cpu/hotplug: provide knobs to control smt
Message-id: <1533921961-30116-19-git-send-email-crecklin@redhat.com>
Patchwork-id: 1732
O-Subject: [RH Confidential: Omega] [RHEL 7.6 CVE-2018-3620 BZ1593384 v5 18/68] cpu/hotplug: Provide knobs to control SMT
Bugzilla: 1593384
CVE: CVE-2018-3620
RH-Acked-by: Jan Stancek <jstancek@redhat.com>
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>

RHEL conflicts:
- RHEL doesn't have /sys/devices/system/cpu/hotplug/ and
  /sys/devices/system/cpu/cpuX/hotplug/, so create cpuhp_sysfs_init()
  from scratch.

Provide a command line and a sysfs knob to control SMT.

The command line options are:

'nosmt':       Enumerate secondary threads, but do not online them

'nosmt=force': Ignore secondary threads completely during enumeration
	    via MP table and ACPI/MADT.

The sysfs control file has the following states (read/write):

'on':          SMT is enabled. Secondary threads can be freely onlined
'off':         SMT is disabled. Secondary threads, even if enumerated
	    cannot be onlined
'forceoff':    SMT is permanentely disabled. Writes to the control
	    file are rejected.

The command line option 'nosmt' sets the sysfs control to 'off'. This
can be changed to 'on' to reenable SMT during runtime.

The command line option 'nosmt=force' sets the sysfs control to
'forceoff'. This cannot be changed during runtime.

When SMT is 'on' and the control file is changed to 'off' then all online
secondary threads are offlined and attempts to online a secondary thread
later on are rejected.

When SMT is 'off' and the control file is changed to 'on' then secondary
threads can be onlined again. The 'off' -> 'on' transition does not
automatically online the secondary threads.

When the control file is set to 'forceoff', the behaviour is the same as
setting it to 'off', but the operation is irreversible and later writes to
the control file are rejected.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>