[x86] kvm: drop l1tf msr list approach
Message-id: <1533921961-30116-53-git-send-email-crecklin@redhat.com>
Patchwork-id: 1763
O-Subject: [RH Confidential: Omega] [RHEL 7.6 CVE-2018-3620 BZ1593384 v5 52/68] x86/kvm: Drop L1TF MSR list approach
Bugzilla: 1593384
CVE: CVE-2018-3620
RH-Acked-by: Jan Stancek <jstancek@redhat.com>
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>

The VMX module parameter to control the L1D flush should become
writeable.

The MSR list is set up at VM init per guest VCPU, but the run time
switching is based on a static key which is global. Toggling the MSR list
at run time might be feasible, but for now drop this optimization and use
the regular MSR write to make run-time switching possible.

The default mitigation is the conditional flush anyway, so for extra
paranoid setups this will add some small overhead, but the extra code
executed is in the noise compared to the flush itself.

Aside of that the EPT disabled case is not handled correctly at the moment
and the MSR list magic is in the way for fixing that as well.

If it's really providing a significant advantage, then this needs to be
revisited after the code is correct and the control is writable.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>