[kernel] cpu/hotplug: set cpu_smt_not_supported early
Message-id: <1533921961-30116-60-git-send-email-crecklin@redhat.com>
Patchwork-id: 1772
O-Subject: [RH Confidential: Omega] [RHEL 7.6 CVE-2018-3620 BZ1593384 v5 59/68] cpu/hotplug: Set CPU_SMT_NOT_SUPPORTED early
Bugzilla: 1593384
CVE: CVE-2018-3620
RH-Acked-by: Jan Stancek <jstancek@redhat.com>
RH-Acked-by: Josh Poimboeuf <jpoimboe@redhat.com>

The CPU_SMT_NOT_SUPPORTED state is set (if the processor does not support
SMT) when the sysfs SMT control file is initialized.

That was fine so far as this was only required to make the output of the
control file correct and to prevent writes in that case.

With the upcoming l1tf command line parameter, this needs to be set up
before the L1TF mitigation selection and command line parsing happens.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>