[netdrv] ibmvnic: Remove code to request error information
Message-id: <20180810173200.2936.41491.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 226633
O-Subject: [PATCH RHEL7.6 BZ1614652 1/2] ibmvnic: Remove code to request error information
Bugzilla: 1614652
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1614652

Description:
------------
When backing device firmware reports an error, it provides an
error ID, which is meant to be queried for more detailed error
information. Currently, however, an error ID is not provided by
the Virtual I/O server and there are not any plans to do so. For
now, it is always unfilled or zero, so request_error_information
will never be called.  Remove it.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

Upstream Status:
----------------
commit 79dabbb7161f43ed29ca888a2488f59e47171aee

https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git/commit/?id=79dabbb7161f43ed29ca888a2488f59e47171aee

----------------
Steve Best

Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>