[cpufreq] Fix possible circular locking dependency
Message-id: <1534875945-26050-1-git-send-email-longman@redhat.com>
Patchwork-id: 227379
O-Subject: [RHEL7.6 PATCH v2] cpufreq: Fix possible circular locking dependency
Bugzilla: 1529668
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1529668
BuildInfo: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=17976435
Upstream: RHEL7-only

The cpufreq driver in RHEL7 is quite different from upstream and it is
not clear if the same problem exists in the upstream code. So this patch
is not sent upstream.

The BZ reports the following possible circular lock dependency:

        CPU0                    CPU1
        ----                    ----
   lock(s_active#95);
                                lock(&policy->rwsem);
                                lock(s_active#95);
   lock(&policy->rwsem);

  *** DEADLOCK ***

 Call Trace:
  [<ffffffffb91cd88c>] dump_stack+0x19/0x1b
  [<ffffffffb91c6cad>] print_circular_bug+0x1f9/0x207
  [<ffffffffb8b2a517>] check_prevs_add+0x957/0x960
  [<ffffffffb8b2bcfc>] __lock_acquire+0xdac/0x15f0
  [<ffffffffb8b2ce29>] lock_acquire+0x99/0x1e0
  [<ffffffffb8fe8bc6>] ? show+0x46/0x90
  [<ffffffffb91d3eb1>] down_read+0x51/0xb0
  [<ffffffffb8fe8bc6>] ? show+0x46/0x90
  [<ffffffffb8fe8bc6>] show+0x46/0x90
  [<ffffffffb8cff7cc>] sysfs_read_file+0xac/0x1b0
  [<ffffffffb8c6997f>] vfs_read+0x9f/0x170
  [<ffffffffb8c6a89a>] SyS_read+0x8a/0xf0
  [<ffffffffb91e2ec9>] system_call_fastpath+0x16/0x1b

The easiest way to prevent this possible circular lock dependency from
happening is to use a trylock to acquire the policy->rwsem in the show
method as that has the minimal impact. If the trylock fails, a EBUSY
error will be returned to the user.

Signed-off-by: Waiman Long <longman@redhat.com>
Signed-off-by: Bruno E. O. Meneguele <bmeneg@redhat.com>